refactor: moved mocks to one folder
This fixes: #46
