+++ test/application/services/character.service.test.ts
@@ -1,7 +1,7 @@
 import {Test} from "@nestjs/testing";
 import {CharacterService} from "../../../src/application/services/character.service";
 import CharacterEntityMapper from "../../../src/infrastructure/mappers/character-entity.mapper";
-import CharacterRepositoryMock from "../../infrastructure/mocks/character.repository.mock";
+import CharacterRepositoryMock from "../../mocks/character.repository.mock";
 import {Repository} from "typeorm";
 import CharacterEntity from "../../../src/infrastructure/entities/character.entity";
 import {getRepositoryToken} from "@nestjs/typeorm";
