+++ test/mocks/character.service.mock.ts
@@ -1,10 +1,10 @@
-import {CharacterService} from "../../../src/application/services/character.service";
-import Character from "../../../src/domain/models/character";
-import CharacterCreationArgument from "../../../src/application/arguments/character/character-creation.argument";
-import CharacterCreationDto from "../../../src/application/dtos/character/character-creation.dto";
-import CharacterUpdateArgument from "../../../src/application/arguments/character/character-update.argument";
-import CharacterUpdateDto from "../../../src/application/dtos/character/character-update.dto";
-import CrudService from "../../../src/domain/interfaces/crud-service.interface";
+import {CharacterService} from "../../src/application/services/character.service";
+import Character from "../../src/domain/models/character";
+import CharacterCreationArgument from "../../src/application/arguments/character/character-creation.argument";
+import CharacterCreationDto from "../../src/application/dtos/character/character-creation.dto";
+import CharacterUpdateArgument from "../../src/application/arguments/character/character-update.argument";
+import CharacterUpdateDto from "../../src/application/dtos/character/character-update.dto";
+import CrudService from "../../src/domain/interfaces/crud-service.interface";
 
 class CharacterServiceMock implements CrudService<Character> {
     findAll(): Promise<Character[]> {
