+++ test/mocks/character.use-case.mock.ts
@@ -1,11 +1,11 @@
-import Character from "../../../src/domain/models/character";
-import Uuid from "../../../src/domain/value-objects/uuid";
-import CharacterUseCase from "../../../src/application/use-cases/character.use-case";
-import CharacterCreationArgument from "../../../src/application/arguments/character/character-creation.argument";
-import CharacterUpdateArgument from "../../../src/application/arguments/character/character-update.argument";
-import CharacterUpdateDto from "../../../src/application/dtos/character/character-update.dto";
-import CharacterCreationDto from "../../../src/application/dtos/character/character-creation.dto";
-import CrudUseCase from "../../../src/domain/interfaces/crud-use-case.interface";
+import Character from "../../src/domain/models/character";
+import Uuid from "../../src/domain/value-objects/uuid";
+import CharacterUseCase from "../../src/application/use-cases/character.use-case";
+import CharacterCreationArgument from "../../src/application/arguments/character/character-creation.argument";
+import CharacterUpdateArgument from "../../src/application/arguments/character/character-update.argument";
+import CharacterUpdateDto from "../../src/application/dtos/character/character-update.dto";
+import CharacterCreationDto from "../../src/application/dtos/character/character-creation.dto";
+import CrudUseCase from "../../src/domain/interfaces/crud-use-case.interface";
 
 class CharacterUseCaseMock implements CrudUseCase<Character>{
     findAll(): Promise<Character[]> {
